+++ database/migrations/2020_05_21_200000_create_team_user_table.php
@@ -4,7 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateTeamUserTable extends Migration
+return new class extends Migration
 {
     /**
      * Run the migrations.
@@ -33,4 +33,4 @@ class CreateTeamUserTable extends Migration
     {
         Schema::dropIfExists('team_user');
     }
-}
+};