+++ resources/views/components/dialog-modal.blade.php
@@ -11,7 +11,7 @@
         </div>
     </div>
 
-    <div class="px-6 py-4 bg-gray-100 text-right">
+    <div class="flex flex-row justify-end px-6 py-4 bg-gray-100 text-right">
         {{ $footer }}
     </div>
 </x-jet-modal>